requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
gunicorn
httpx>=0.27
psycopg[binary]